requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
certifi
click
mongoengine
pillow
pip
pymongo
python-dotenv
python-settings
random-object-id
setuptools
vc
wheel
wincertstore